set the current working directory in the chroot
This patch allows you to set the current working directory (in the
chroot) before running a command with --chroot.  This avoids the need to
pass shell snippets ('cd /some/path && /run/cmd') to mock when running a
command that expects to executed from a certain directory.  It's useful
when using --copyin to setup the environment before running a command.

>From e4071d1d41a62ccf4461dfab958f9325edf30c97 Mon Sep 17 00:00:00 2001
From: Mike Bonnet <mikeb@redhat.com>
Date: Thu, 24 Jan 2008 17:09:06 -0500
Subject: [PATCH] optionally set the current working directory (in the chroot) before running command with --chroot

Signed-off-by: Michael E Brown <mebrown@michaels-house.net>