Support configurable root cache compression
This patch adds configuration options that influence how is the root
cache compressed:

  config_opts['plugin_conf']['root_cache_opts']['compress_program'] = "gzip"
  config_opts['plugin_conf']['root_cache_opts']['extension'] = ".gz"

Motivation to do this was to be able to use lzma compression. While it
takes a long time to compress, for me it was more than two times faster
than gzip when decompressing. Since I only seldom regenerate the root cache
while doing a lot of builds, I was able to gain a considerable speed up.