Retain order of umountCmds
The order of execution of umounts is the reverse of the order of
the mounts, so as to deal with a hierarchy of mounts correctly.
The ordering is reversed by _umountall but was being left in the
reversed state so that the next time _umountall was called, it
attempted to unmount filesystems in the wrong order, leading to
build failures due to being unable to unmount for example /proc
whilst /proc/filesystems was still mounted. This change reverses
the order of umount commands back again before _umountall exits
so as to maintain the ordering.

Signed-off-by: Clark Williams <williams@redhat.com>