add runtime location of plugins
Mock has plug-in mechanism but source code modification is needed to add a
plug-in. For writing an experimental plug-in or a plug-in for private
purpose, it is nice if one can add a plug-in without modifying the source
code of mock itself.

With this commit, python script files at $PKGPYTHONDIR/plugins are
added as plugins to mock automatically.

Signed-off-by: Clark Williams <williams@redhat.com>