fix chroot cleanup issues (BZ# 668222)
The cleanup_on_success and cleanup_on_failure options suffer from 2 issues.

First, if the options are specified in the configuration file and no resultdir
was specified on the CLI, then those options will be forced to False (even if
the resultdir is specified in the configuration file).

Second, if the resultdir is specified but is a subdirectory of the build root,
then we must not clean it, even though both are different.

So the appropriate thing to do is to check if the real resultdir (not the one
from the CLI parameters) is a subdirectory of the build root, which covers both
issues.

Signed-off-by: Clark Williams <williams@redhat.com>