fix problem in scm.py with SSH_AUTH_SOCK [BZ# 803217]
Make a copy of the environment for the SCM plugin and add
the SSH_AUTH_SOCK to it (as well as changing HOME to match the
homedir of the active pid).  Use the SCM environment when running
subprocesses for SCM operations.

Add SSH_AUTH_SOCK to the KEEP_ENV_VARS config in consolehelper
so that SCM operations work properly.

Signed-off-by: Clark Williams <williams@redhat.com>