reworked the state mechanism to have start/end and allow nesting
Reworked the chroot state to be a stack of states. Added the chroot
methods start() and finish() to indicate start and end of a state.
Modified the state() method to return the current state at the top
of the stack (no modificaiton of stack). Added the alldone() method
to check for unfinished states.

Signed-off-by: Clark Williams <williams@redhat.com>

Conflicts:

	py/mockbuild/backend.py