Fix mock kernel version comparison [BZ# 847473]
Mock's comparison of kernel versions fails when the version has
less than 3 version components (e.g. 3.4).

This commit fixes the problem by calling the proper yum function,
compareVerOnly, since this is just a version comparison, not an RPM
EVR comparison. I also renamed the function to cmpKernelVer, although
that can be backed out in the name of API compatibility.

Signed-off-by: Clark Williams <williams@redhat.com>