Pass values of --plugin-option through literal_eval [BZ# 1018359]
The value of --plugin-option is set to be of type string when adding
the option to the OptionParser. The values are then passed as string
to the respective option receivers. This makes it impossible to pass
i.e. bind_mount:dirs via --plugin-option, since bind_mount.py expects
dirs to be an 2d-iterable (i.e. tuple of tuples), but it will get a
string.

One possibly clean fix would be to pass all values of --plugin-option
though literal_eval. If literal_eval fails, then it is likely an actual
string, and it is kept as such. Otherwise, the string is parsed to a
python object.  The only risk is that the user passes an argument which
should be a string but also is valid python syntax. To avoid these
issues, the user should quote the value, i.e.
	--plugin-option="myplugin:key='value'"

Signed-off-by: Clark Williams <clark.williams@gmail.com>