Add 'extra_chroot_dirs' config option
Add the 'extra_chroot_dirs configuration option to allow creation
of directories other than the core set inside the chroot. This is
done by setting the variable in a config file:

   config_opts['extra_chroot_dirs'] = [ '/tmp/foo', '/tmp/bar' ]

This will result in the dirs /tmp/foo and /tmp/bar being created
in the chroot when it is initialized.

Signed-off-by: Clark Williams <clark.williams@gmail.com>