Don't output installation/build output when redirected
The installation and build output goes to stdout and to root.log/build.log. If
the output is redirected to a file, it means that no one is watching it in
realtime and there's no reason to duplicate it across two files. There are
multiple applications that capture mock's output (Koji, fedora-review). Because
some build logs can be tens of megabytes it is necessary that the duplication
is prevented.