With --no-clean, delete /builddir except build/SOURCES [RHBZ#1173126]
In order to be able to have the ability to execute more successive builds with
--no-clean, it's necessary to cleanup build directories from the previous
build. Now the whole /builddir is deleted except SOURCES directory which is
retianed to support nosrc builds. This is further configurable with new config
option 'exclude_from_homedir_cleanup', which contains a list of paths relative
to /builddir that will be excluded from deletion in prebuild.