mounts: don't try and mount /dev/shm or /dev/pts if internal setup false
If the config 'internal_dev_setup' is False (default is True) then do not
add /dev/shm and /dev/pts to the mount list. The mount will fail because
their mount directories will exist (created in _setup_devices()). Usually
the setup is done by bind-mounting the system /dev, but if that's not done
then the external setup will be responsible for creating appropriate
mountpoints.

Signed-off-by: Clark Williams <clark.williams@gmail.com>