Add pm_request plugin
pm_request plugin listens to requests for package management commands from
within the buildroot, using unix socket. It can be used by buildsystems, such
as Maven, to automatically install missing packages if they're available. It's
disabled by default to not affect build reproducibility. A warning is issued at
the end of the build if the plugin was used.