Add "rpmbuild_networking" key (False by default) for nspawn backend
Build systems like Koji and many others want to support one aspect of
reproducible builds by ensuring all input artifacts are cached within
the organization.  This is the "If Github goes down, you can still
build" aspect.

Traditionally, Koji installations implement this via firewalling
rules on the server.

However, now with the systemd nspawn backend, we can rely
on much more robust container features.  In particular, systemd has
a `--private-network` option which does exactly what we want.

Teach mock how to use it, and enable this by default if the nspawn
backend is in use.

This has the potential to break mock users who were relying on
networking - but it's easy to turn back on.  But I assert that the
majority of people using mock are exactly the kind of user who want to
do full packaging, and it's best if they find out about any network
issues before they do a Koji/whatever build.