Add config for CentOS PPC64LE arch
CentOS for PPC64LE is a altarch SIG announced in the end of 2015[1] and this
patch provides configuration file to use mock on it.

Note that there are two GPG keys for the PPC64LE repositories at
mirrors.centos.org, in general noarch packages were signed with the usual
RPM-GPG-KEY-CentOS-7 and packages rebuilt or created for the architecture, in
the case PPC64LE, were signed with RPM-GPG-KEY-CentOS-SIG-AltArch-7-ppc64le[2],
also added in this patch.

[1]: https://lists.centos.org/pipermail/centos-announce/2015-December/021556.html
[2]: https://www.centos.org/keys/