systemd-nspawn: run as PID2 #36
Since we don't run systemd or any process which implements sysvinit
compatible signal handling (specifically: it needs to reboot on SIGINT,
reexecute on SIGTERM, reload configuration on SIGHUP, and so on) we
should use -a/--as-pid2 to run STUBINIT as PID1 and our process as PID2.

rpmbuild doesn't implement sysvinit compatible signal handling (e.g.
it doesn't kill child processes, so builds can stuck because of this).

Reported-by: Vít Ondruch vondruch@redhat.com
References: https://bugzilla.redhat.com/show_bug.cgi?id=1372234
Signed-off-by: Igor Gnatenko ignatenko@redhat.com
Edit to work on EL by msuchy