change_home_dir: Actually set ownership of homedir
Previously only the ownership of subdirectories and files in the
homedir were set, not that of the homedir itself.

We now also use uid_manager.changeOwner() instead of re-implementing
it. This has the side effect of no longer doing chmod on everything,
but there should really not be any reason for that.