Buildroot: Ensure homedir and build dirs always have correct owner
When running mock --no-clean results in mock having to start with
unpacking a fresh root_cache permissions were not properly set.
In this case we avoid doing a recursive chown, and only change
ownership of the directories that really must be owned by the current
mock user for anything at all to work.