Bootstrap package manager using outer chroot
This commit changes Mock so that it first builds
an outer chroot, then uses that chroot to build the main
chroot. This means the chroot is built with the version of
dnf or yum it contains, allowing the package being built to
use newer RPM/dnf features. This means the only packages
that need to be installable by the old rpm/yum/dnf are those
required in this minimal environment.