create /dev nodes even when using nspawn [RHBZ#1467299]
This happen with initial population of chroot. I.e the /dev/null initialy exists, but when postscript is run, then because dnf is executed with --installroot and rpm is run inside the chroot. But /dev is not mounted there so /dev/null really does not exists yet. And we cannot rely on nspawn to populate because this is initial command which is not executed within chroot, but from host, because there is no binary installed in chroot yet.