Update Python 2 dependency declarations to new packaging standards
According to Fedora Packaging guidelines for Python, packages must use names with either python2- or python3- prefix in requirements where available.
We are aiming to rename python-* dependencies to python2-*, so we can later switch the python-* namespace to Python 3.

This PR is part of Fedora's Switch to Python 3 effort.

This has been taken from:
https://src.fedoraproject.org/rpms/mock/pull-request/3#request_diff

This is safe as RHEL7 python-devel provides python2-devel too.