call 'gtar' binary, instead of 'tar'
This helps OpenMandriva where /bin/tar is equivalent for
/bin/bsdtar.  Such setup is somewhat unexpected on GNU/Linux
distro, but anyway - the fix doesn't hurt Fedora/RHEL/CentOS where
we symlink gtar to tar since ever.

As original issue suggests, we could implement everything using
the POSIX tar options (+ helper scripting); but that would be just
too much work which is not worth wasting ATM, at least as long as
it is safe to expect that 'gtar' is in PATH and that it really is
GNU tar.

Resolves: #169