completion: fix completion after options without arguments
The completion script contained a list of many options which had no
completion available (some required an argument, others did not).  The
issue with this is that no completion was available following the use of
one of these options.  For example, 'mock --nocheck --<TAB>' provides no
completion.

Remove the explicit 'return 0' for all options which do not require an
argument.  Only return when the previous option is expected to be the
last (or only) option used, e.g. -h, --help, --version; or when the
option requires an argument for which we do not or cannot offer
completion.  Keep these two cases separate and note the reason each case
exists to make it clear when new options are added.