When installing into the buildroot, ignore /proc and /sys
If mock is running in a user namespace, the the /proc and /sys we
mount into the buildroot will be owned by nobody:nobody (the real
host on the root) not root:root. To avoid failures during package
installation set the %_netsharedpath RPM macro to /proc:/sys to skip
these directories. (Suggestion from Panu Matilainen.)

Implement by pointing HOME to a directory with only ~/.rpmmacros.
(This means that /root/.rpmmacros will now be ignored; system
RPM configuration will still be honored.)