Use a bind mount for /sys instead of remounting it
Mounting a fresh sysfs from within a user namespace is allowed only
in limited cases, so always use a bind mount instead to avoid problems.

This needs to be a recursive bind mount so that any mounts on top
of the parent /sys are preserved in the child (the kernel will
fail a plain bind mount).