Fix mock for non-ascii paths on python2
Many paths in mock are a combination of a directory sepcified in
configuration with a subdirectory read in from the filesystem (or
specified on the commandline or from some other system service).  The
configuration paths were made into text strings by the Jinja2 templated
configs feature.  On python2 (el6 and el7) the strings from the
filesystem are byte strings.  When the two string types are combined, a
UnicodeError is thrown.

Rather than find all the places where mock is loading byte strings and
would need to transform them back into text strings, this patch merely
changes the text strings returned from Jinja2 into native strings (ie:
text strings on python3 and byte strings on python2).  This should
restore the behaviour of the pre-Jinja2 configuration.