Allow AArch64 systems to build 32-bit ARM packages
There are a number of AArch64 systems that natively support
building and running code for 32-bit ARM ISAs.

In particular, some of the new ARM cloud servers are capable of it,
like the new ARM cloud servers from Packet.com.

RPM already has enhancements for supporting this as well.

So, let's go ahead and allow it.