grant anyone access to bind-mounted /etc/resolv.conf
The mkstemp() call creates file with 0600 permissions, and it is
not a natural setup for the /etc/resolv.conf file.

The 0600 permissions are especially problematic with
'systemd-nspawn --private-users=pick' (recommended settings by
systemd docs, and used e.g. by copr in some cases).  The option
doesn't imply that the file is re-mapped according to the picked
user namespace.

Fixes: #265