Make the mount namespace reliably private
The code attempted to do all mock mounts inside a private mount namespace,
by using unshare(CLONE_NEWNS), but it turns out that that isn't effective
if part or all of the original mount tree are marked as 'shared' mounts -
changes to copies of those mounts will still propagate back to the original
mounts. Mark the entire new mount tree as 'private' to avoid such sharing.