DynamicBuildrequires: Detect when no new packages were installed
Prior to this change, the DynamicBuildrequires loop never ended.

This commit fixes it in 2 ways:

 1. It now actually stops when max_loops (10) happened.
 2. It also stops when the package manager hasn't installed anything.

Both is consistent with the comment present in the previously broken code.

The detection of 2. is based on packages returned by `rpm -qa` before
and after the installation. If they haven't changed, the package
manager hasn't installed anything.