mock: only make /sys and /proc mounts rprivate
All the mounts in mock are executed after unshare() call.  Marking the
root mount "rprivate", even though it looked to be safe, entirely broke
the LVM and tmpfs plugins, see the issue #311.  So let's drop the global
--make-rprivate at least for now (#317).

The '--make-rprivate' was used before because we needed to
protect/isolate host's mount table, because in PR#236 we started to
(bind)mount the /<buildchroot>/{proc/sys} recursively (even though we
actually didn't need "recursive" mount from mock perspective - at least
assuming that everything worked fine before PR#236 with normal mount).
The reason for --rbind was rhbz#1745048.

The consequence of recursive mount was that /sys/fs/selinux suddenly
become available in chroot, and it made the software in chroot think
that selinux was working there (because /usr/sbin/selinuxenabled wasn't
telling truth).  So we tried to enable selinux module everywhere to fix
that (commit 964235acf108), but it wasn't enough -- the plugin itself
did not yet affect the /sys/fs/selinux mount point.  Fixing this now,
the plugin overmounts /sys/fs/selinux.

Even though we are dropping '--make-rprivate /', we still have to have
the /sys mounted recursively, and we have to make sure that
lazy-umounting that won't affect host.  So we still need to make /sys
mount 'rprivate'.

The lazy umount of recursively bind-mounted /proc had similar problems
as /sys (rhbz#1744996), so ditto for /proc.

Btw., as a side effect, this commit fixes #1740377 by removing the
--make-rprivate call.

Reverts: 377fa52c61654dc75bed78933b40c343a692615c
Resolves: rhbz#1740421, rhbz#1740377, rhbz#1744996
Fixes: #311, PR#315