mock: don't create root files if possible
We don't want to create root-owned dirs/files as done in a239b59b871c8c,
so we revert part of the commit.  But the `--chain` option requires us
to create the temporary directory as non-privileged user, so mock can
later store files/create dirs underneath.

While we are on it, the remaining root-owned leftover would be the
repodata directory in local_repo_dir, so make sure that createrepo is
run as non-privileged user as well.

This reverts commit 92aaabe44824bb2bd95792dcf4b80767ba30d24a, it is not
needed now.

Fixes: #341, #322 (corrected fix), rhbz#1751944