when scrubbing all, remove also bootstrap container
Fix #351

The `Commands.scrub` method requires `bootstrap_chroot` object initialized
in order to work with it. I.e. it doesn't scrub a bootstrap chroot unless
the `--bootstrap-chroot` parameter is set. Creating its own object inside
the method would duplicate a lot of existing code and doing it without the
object and its methods and properties is IMHO not a good idea.

Since the workaround for the #351 is

    mock -r <chroot> --bootstrap-chroot --scrub=all

we can just simply pretend that `--bootstrap-chroot` was set,
when using `--scrub`.