mock: fix broken rpmbuildstate
Previously, when something in 'rebuild_package()' method threw an
expection the state.finish(rpmbuildstate) wasn't called, and this caused
another exception in finally block where we tried to
state.finish(buildstate):

  mockbuild.exception.StateError: state finish mismatch: current:
  rpmbuild dummy-pkg-20191008_0834-0.src.rpm, state: build phase for
  dummy-pkg-20191008_0834-0.src.rpm

This exception was not handled at all, so uid_manager didn't have a
chance to raise privileges again by restorePrivs() - and the exception
bubbled up to main().  In the meantime, the cleanup_on_failure=True
handler was executed in rebuild_generic() - but without administrator
privileges so it failed.

Fixes: #349