mock: drop two unnecessary privilege escalations
At those points in code path, mock already had appropriate permissions.
So I believe this was a workaround for bug fixed by previous commit.

More, one of the privilege escalation calls was not taken back when
bootstrap_chroot was disabled and mock seemingly worked, so it sort of
confirms that this commit isn't incorrect.