bootstrap: fix --install and --installdeps
To fix --install* options for bootstrap transparently, all the files
given on command-line need to be bind-mounted into bootstrap chroot so
they can be processed _inside_.

Since the files on command-line may be relative to cwd and we don't want
to create arbitrary directory structures within bootstrap buildroot, we
simply bind-mount the files into "home" directory.  The obvious
consequence is that multiple files with the same basename could clash
inside in the homedir, so we rather raise an error as soon as possible
instead of overmounting the same path.

Fixes: rhbz#1447627