bootstrap: assure /bin/yum is from dnf-yum on el8
This is likely a bug caused by the fact that `--yum` option tells:
1. to use /bin/yum on host to install bootstrap
2. to install /bin/yum inside bootstrap to install normal chroot.

The problem of 1. is that a) in el8-bootstrap is that yum.rpm is not
easily installable (see #233) and b) that user typically doesn't want to
use /bin/yum in bootstrap to install dnf-capable system.

So 1. and 2. are two orthogonal things, and we should have two config
options for this (or use --yum only to affect host pkg manager
decision).

Fixes: #233