sign-plugin: use %(rpms) variable expansion again
It is non-trivial to assure that {{ rpms }} specified in mock
configuration isn't expanded (to empty string) too early.  Proper
solution to that will take some time, so for now revert back to the
%(rpms) syntax.

We should stay compatible with %(rpms) if we decided to move to
{{ rpms }} again in future.

Add integration test for rpmsign, plus preparation script.

This reverts commit d44ec018b59822c8ba6dcffb0002148b57a69457.

Fixes: #476, rhbz#1806577