mock: don't clean between consecutive builds
Fixes: #483

The issue doesn't seem to be --chain specific. It is reproducible by both

    mock -r <chroot> --chain foo.src.rpm bar.src.rpm
    mock -r <chroot> foo.src.rpm bar.src.rpm --resultdir=/foobar

Unfortunatelly the fix can't be done in some common code, but separately
for both of these cases. Basically we need to make sure to not call
`commands.clean()` between consecutive builds. Interesting is, that
if --chain is not used, we can clean before the first build and still
go from cache, but when using --chain, we can't clean before any of them.