Better support for bootstrap_package_manager config option
The bootstrap buildroot is constructed before mapping bootstrap_*
options into config options for the bootstrap buildroot, but the
constructor of the buildroot wants to know what the desired package
manager to set up the bootstrap buildroot is. If a
bootstrap_package_manager option is specified, this should be used
as the desired package manager rather than what's been requested for
setup of the actual main buildroot.

This is not a terribly important option to get right since mock will
check what's available on the host system and fall back sensibly.
However, this will usually result in a warning message appearing in
the logs, which is nice to avoid.