tests: test building mock.src.rpm only against certain configs
The mock.spec isn't compatible with old chroots (like epel-6), and with
opensuse/mandriva chroots (build dependencies missing).  This could be
fixed (compat patches are welcome), but it isn't burning issue ATM.

So stop testing all the configs with builds of mock.src.rpm when it can
not work, and build test-C.src.rpm instead.