mock: implement Buildroot.doOutChroot() and fix package_state
With this, when bootstrap is enabled, we use compatible RPM
implementation with the rpmdb generated in destination chroot.

Currently for example - the new method fixes package_state plugin for
OpenMandriva chroot built on Fedora host (incompatible rpmdb backends).

But this is much more important fix for the upcoming builds of
Fedora Rawhide packages; RPM on F33+ is going to have SQLite backend by
default:
https://fedoraproject.org/wiki/Changes/Sqlite_Rpmdb

Resolves: #525