mock: conflict with old core-configs
We want to make the mock-core-configs package "optional", in a sense
that users may replace it with their own package.  This should be
acceptable use of Conflicts tag:
https://docs.fedoraproject.org/en-US/packaging-guidelines/Conflicts/#_optional_functionality

Keep the Requires tag -- as probably all users expect to have mock
configured by default.  And also to not cause wave of updates in all the
packages that depend on mock.  To ease the `mock-core-configs`
replacement, flip the 'Requires:' against new explicit 'Provides:'
from mock-core-configs that we don't Conflict with.

Fixes: #544