mock: unbind mock from mock-core-configs package
Newly we "only" conflict with too old mock-core-configs packages.

We want to make the mock-core-configs package "optional", in a sense
that users may replace it with their own package.  This should be
acceptable use of Conflicts tag:
https://docs.fedoraproject.org/en-US/packaging-guidelines/Conflicts/#_optional_functionality

But keep the Requires tag (a) as probably all users expect to have mock
configured by default and (b) to not cause wave of updates in all the
packages that depend on mock+configs.  To ease the `mock-core-configs`
replacement, flip the 'Requires:' against new explicit 'Provides:' from
mock-core-configs that we don't Conflict with.  Still, mock Suggests
the default mock-core-configs as a preferred variant.

Fixes: #544