bootstrap: provide predictable resultdir
The problem in GH#539 is that _resetLogging() is called twice initially
(for both bootstrap and target chroot) so it always initially
(force=False) adds two FileHandlers to the _same_ logger objects.  And
bad things happened when both the handlers were instructed to write to
the same file.

It's not ideal time to resolve the logging mess properly (that will be
large and risky patch), so I have at least work-around that should
minimize the logging problems.

Related: #539