mock: move site-defaults.cfg from /etc to %doc
The file was not pre-configured by default (every option there was
commented-out - we specify defaults in util.py) so it always existed for
documentation purposes anyway.

We keep the full site-defaults.cfg file in mock.rpm as %doc, but now we
can move the site-defaults.cfg file to mock-core-configs.rpm so it is
shipped with other %config files.

Fixes: #546