mock: re-add "Installed packages:" info after --init
This has been broken with bootstrap for some time, but because it was
always called with raiseExc=False we did not notice it (the only effect
was that the stdout output was slightly shorter).

Let's replace util.do() with buildroot.doOutChroot to transparently use
rpm from bootstrap chroot instead of the host's one.

I also dropped the uid/gid/env overrides as those shouldn't be needed
(we can calmly run as root here).

Fixes: #565