Better error for dynamic_buildrequires %prep failure
There were two more issues in dynamic_buildrequires handler.  First that
we set success=True anytime `rpmbuild -br` succeeded.  This would exit
the build requires loop too early if `-br` eventually returned exit
status 0.

The second is that we tried to install the nosrc.rpm file in finally
block, which was done everytime the `rpmbuild -br` command failed (not
only when exit status was 0 or 11).

Fixes: #570, #581