mock: copy source CA certificates
Prior to this change, we would only copy the "extracted" SSL CA
certificates into the chroot. If anything ran "update-ca-trust" inside
the chroot, this would delete our custom SSL certificates from the
"extracted" directory. For example, Fedora and RHEL's main
"ca-certificates" package always does this in %post, and any custom
third-party package could do this as well.

Copy the entire parent directory so that "sources" and "extracted" are
both present in the chroot. With this change, "update-ca-trust"
does not wipe out the CA certificates from the chroot.

Fixes: #588