Fix site-defaults.cfg reference to docs
On systems where %_pkgdocdir isn't specified or it points to versioned
directory, the docs directory looks like `../mock-core-configs-32.3/`
and we want to point user to `../mock-2.3/`.  At that build-time moment
we aren't able to tell the version of _mock_, so it is better to point
the user only to shell wild-card `../mock-*/`.

Per report in #597