ELN should use for build Everything repository
ELN BaseOS, AppStream and CRB repositories are for runtime. However, as in RHEL
there are even packages which are just in the buildroot. These packages are not
part of the repositories above but only Everything repository.

So to reflect how the next RHEL will be build we should use the Everything
repository which should contain all the packages not just ones which will be
shipped. And to that we also need to enable AppStream repository otherwise we
won't have modules. Disable everything else.