Fix bootstrapping of newer Fedora on EL7
We forgot to add Fedora 33 host override when adding other Fedora 33
configs.  This is something which would definitely repeat in future, so
I'm rather using a script that will do this hack for us automatically in
the future.

While I'm on it I'm removing rhel-8 override without replacement because
in the meantime the RPM in EL8 got the libzstd support.

Resolves: rhbz#1914750