mock: Work around bug setting propagation for recursive bind-mounts
In util-linux versions before 2.33, the mount program does not make
new bind-mounts recursive when the propagation is specified in the
same command (rhbz#1584443). This affects RHEL 7 and 8.

Instead, make the bind-mount first, then remount it to change mount
options and/or propagation. This is the "classic" method described
in the mount(8) man page, which explains that userspace must issue
separate system calls, even if the options are provided in a single
command. Use this method for all bind-mounts.

Also, adjust the return value from this method when the bind-mount
is already mounted, for consistency.