config: Disable Rocky Linux "Devel" repo by default
Having this enabled makes it too easy for people to accidentally
build packages that depend on content in here at runtime. That
makes it possible for people to make packages that will not be
compatible with RHEL and derivatives.

Disable this repo by default. If people need this, they can enable
it in the Mock commandline as-needed. At least then they acknowledge
what they're doing and know the consequences.