mock: don't pass source to `mount -o remount`
The mount program with `-o remount` behaves differently when source is not
passed (rhbz#1985628). Though in a normal host the result is the same no matter
whether source argument is passed, in some special environment, for example in
some Toolbox container, passing source to the mount program results in a
permission error.

If source is passed, the options are passed directly to mount(2):

> # mount -n -o remount,ro,bind /usr/local/ /var/lib/m

results in

> mount("/usr/local", "/var/lib/m", 0x56381a4532d0,
>  MS_RDONLY|MS_REMOUNT|MS_BIND, NULL)

This will fail with permission error in some Toolbox containers, since there are
many options that cannot be reset inside the container.

But if source is not passed, the final arguments passed to mount(2) are combined
with the current mount options.

> # mount -n -o remount,ro,bind --target /var/lib/m

results in

> mount("fuse-overlayfs", "/var/lib/m", 0x55d9a97f3310,
>  MS_RDONLY|MS_NODEV|MS_REMOUNT|MS_NOATIME|MS_BIND,
>  "user_id=0,group_id=0,default_per"...)

This works in Toolbox containers.

So we just don't pass the source argumnent and the `mount -o remount` command
will work in hosts and containers

Closes: #715