mock: don't specify SOURCE when remounting bind-mounts
The mount program with `-o remount` behaves differently when source is
not passed (rhbz#1985628).  When mount source is specified, all the
mount options are specified on the command line from scratch.  Whilst
when source isn't specified, mount command combines the set of specified
options with the set of pre-existing mount options of the "old"
mountpoint.

The remount command is more likely to succeed when source isn't
specified as some "old" mount options might be crucial.  For example, in
some restricted environments (e.g. Toolbox) we wouldn't be allowed to
remount a bind-mount without the 'nodev' option.  Remounting read-only
with the 'mount -o remount,bind,ro SRC DEST' command wouldn't succeed
because of the missing 'nodev' option.  Remounting with just
'mount -o remount,bind,ro DEST' succeeds because the 'nodev' is
inherited from the "old" bind-mount (note that the 'nodev' option is on
the "old" bind-mountpoint because the previous command
'mount -r bind SRC DEST' inherited all mount options from the /src
filesystem).

Thanks to Robin Lee for the patch.

Closes: #715
Closes: #759