config: Migrate EPEL 8 configuration to use AlmaLinux 8 as the base OS
CentOS Linux 8 support ends at the end of 2021, so we need to select
a new base OS for the default EPEL configuration.

AlmaLinux satisfies the requirements to replace CentOS Linux for the
EPEL configuration, namely that:

* Alma provides the exact same content as Red Hat Enterprise Linux 8
* Alma releases the same time as RHEL (Alma 8.5 was 2 days after RHEL 8.5)
* AlmaLinux has committed to supporting all RHEL architectures
* AlmaLinux is freely available with no entitlement encumbrances

This allows all users of Mock to (re)build EPEL packages to successfully
do so regardless of host Linux distribution Mock runs on.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1965480