Add epel9 repos to epel9 template
Originally the plan was to launch epel9-next before epel9.  We (the EPEL
Steering Committee) changed course and decided to launch them together.
This change adds the repo definitions that were previously missing from
the template file and renames the epel-next-9-* config files to match
the new {base}+epel pattern.

https://communityblog.fedoraproject.org/epel-9-is-now-available/